fix: el-message z-index, move some component to directory
