+++ src/components/admin/AdminSystemInfo.vue
